@@ -5,14 +5,14 @@ publish_to: "none"
 version: 0.1.0
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   map_launcher:
     path: ../
-  flutter_svg: ^0.19.1
+  flutter_svg: ^0.21.0
 
 dev_dependencies:
   flutter_test:
