@@ -1,11 +1,11 @@
 name: map_launcher
 description: Map Launcher is a flutter plugin to find available maps installed on a device and launch them with a marker or show directions.
-version: 1.1.3
+version: 2.0.0-nullsafety.0
 homepage: https://github.com/mattermoran/map_launcher
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.10.0"
+  sdk: '>=2.12.0-0 <3.0.0'
+  flutter: '>=1.24.0-10.2.pre <2.0.0'
 
 dependencies:
   flutter:
