@@ -4,8 +4,8 @@ import 'package:map_launcher/map_launcher.dart';
 
 class MapsSheet {
   static show({
-    @required BuildContext context,
-    @required Function(AvailableMap map) onMapTap,
+    required BuildContext context,
+    required Function(AvailableMap map) onMapTap,
   }) async {
     final availableMaps = await MapLauncher.installedMaps;
 
